requirements.txt
streamlit
langchain==0.3.25
langchain-community==0.3.25
langchain-groq
langchain-huggingface
langchain-core>=0.3.58
faiss-cpu
sentence-transformers
pypdf
python-dotenv
huggingface-hub